Use an array for OPENAL_LIBNAME on Posix.
Debian, Ubuntu, Fedora and derivatives ship libopenal.so only with the
development package of libopenal, as a link to libopenal.so.1, which is
shipped by the runtime packages.
So, we try to load libopenal.so.1 first on non-macOS POSIX platforms,
and if it fails we try also to load libopenal.so.
This is similar to what GLAD does.